+++ project/src/components/app/app.tsx
@@ -1,12 +1,12 @@
 import { BrowserRouter, Outlet, Route, Routes } from 'react-router-dom';
-import { MainPage } from '../main-page/main-page';
-import { SignIn } from '../sign-in/sign-in';
-import { MyList } from '../my-list/my-list';
-import { Film } from '../film/film';
+import { Main } from '../../pages/main/main';
+import { SignIn } from '../../pages/sign-in/sign-in';
+import { MyList } from '../../pages/my-list/my-list';
+import { Film } from '../../pages/film/film';
 import { AddReview } from '../add-review/add-review';
 import { Player } from '../player/player';
 import { PrivateRoute } from '../private-route/private-route';
-import { NotFoundPage } from '../not-found-page/not-found-page';
+import { NotFoundPage } from '../../pages/not-found-page/not-found-page';
 import { AppRoute } from '../../constants/routs';
 import { AuthorizationStatus } from '../../constants/auth';
 import { Films } from '../../types/film';
@@ -31,7 +31,7 @@ function App({ limit, promoFilm, films }: Props): JSX.Element {
         >
           <Route
             index
-            element={<MainPage limit={limit} promoFilm={promoFilm} films={films} />}
+            element={<Main limit={limit} promoFilm={promoFilm} films={films} />}
           />
           <Route
             path={AppRoute.SignIn}
@@ -40,7 +40,7 @@ function App({ limit, promoFilm, films }: Props): JSX.Element {
           <Route
             path={AppRoute.MyList}
             element={
-              <PrivateRoute authorizationStatus={AuthorizationStatus.Auth}>
+              <PrivateRoute authorizationStatus={AuthorizationStatus.Unknown}>
                 <MyList films={films} />
               </PrivateRoute>
             }
