+++ project/src/pages/sign-in/sign-in.tsx
@@ -1,12 +1,6 @@
-import { Logo } from '../logo/logo';
-import { Films } from '../../types/film';
-import { FilmList } from '../film-list/film-list';
+import { Logo } from '../../components/logo/logo';
 
-type Props = {
-  films: Films
-}
-
-function MyList({ films }: Props): JSX.Element {
+function SignIn(): JSX.Element {
   return (
     <>
       <div className="visually-hidden">
@@ -27,7 +21,8 @@ function MyList({ films }: Props): JSX.Element {
               fill="#FFF9D9" fillOpacity="0.7"
             />
             <path
-              fillRule="evenodd" clipRule="evenodd" d="M27 23.8571V16H23.8571V23.8571H16V27H23.8571H27L27 23.8571Z"
+              fillRule="evenodd" clipRule="evenodd"
+              d="M27 23.8571V16H23.8571V23.8571H16V27H23.8571H27L27 23.8571Z"
               fill="#FFF9D9" fillOpacity="0.7"
             />
             <path
@@ -68,25 +63,32 @@ function MyList({ films }: Props): JSX.Element {
         <header className="page-header user-page__head">
           <Logo />
 
-          <h1 className="page-title user-page__title">My list</h1>
-
-          <ul className="user-block">
-            <li className="user-block__item">
-              <div className="user-block__avatar">
-                <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-              </div>
-            </li>
-            <li className="user-block__item">
-              <a className="user-block__link">Sign out</a>
-            </li>
-          </ul>
+          <h1 className="page-title user-page__title">Sign in</h1>
         </header>
 
-        <section className="catalog">
-          <h2 className="catalog__title visually-hidden">Catalog</h2>
-
-          <FilmList films={films} />
-        </section>
+        <div className="sign-in user-page__content">
+          <form action="#" className="sign-in__form">
+            <div className="sign-in__fields">
+              <div className="sign-in__field">
+                <input
+                  className="sign-in__input" type="email" placeholder="Email address" name="user-email"
+                  id="user-email"
+                />
+                <label className="sign-in__label visually-hidden" htmlFor="user-email">Email address</label>
+              </div>
+              <div className="sign-in__field">
+                <input
+                  className="sign-in__input" type="password" placeholder="Password" name="user-password"
+                  id="user-password"
+                />
+                <label className="sign-in__label visually-hidden" htmlFor="user-password">Password</label>
+              </div>
+            </div>
+            <div className="sign-in__submit">
+              <button className="sign-in__btn" type="submit">Sign in</button>
+            </div>
+          </form>
+        </div>
 
         <footer className="page-footer">
           <Logo isLight />
@@ -100,4 +102,4 @@ function MyList({ films }: Props): JSX.Element {
   );
 }
 
-export { MyList };
+export { SignIn };
